\name{plot.envelopeJ}
\alias{plot.envelopeJ}
\title{
Plot method.
}
\description{
Plot method.
}
\usage{
\method{plot}{envelopeJ}(x,cex.legend=1,...)
}
\arguments{
  \item{x}{
an envelope object.
}
\item{cex.legend}{size of legend.}
\item{...}{optional.}
}
\author{
Mehdi Moradi \email{m2.moradi@yahoo.com}
}
