\name{rSSI.lpp}
\alias{rSSI.lpp}
\title{
Simulate Simple Sequential Inhibition on a Linear Network}
\description{
Simulate Simple Sequential Inhibition on a Linear Network.
}
\usage{
rSSI.lpp(n, r, L, giveup = 1000, nsim = 1, ...)
}
\arguments{
  \item{n}{
number of points.
}
  \item{r}{
inhibition distance.
}
  \item{L}{
linear network.
}
  \item{giveup}{
number of rejected proposals.
}
  \item{nsim}{
number of simulation.
}
}
\author{
Mehdi Moradi <m2.moradi@yahoo.com>
}
